sheep/nvmet: compilation fixes
Remove 'utils.h' (already present in sheepdog sources) and move the
sql statements into the functions to enable argument checking.

+++ sheep/nvmet/ops.h
@@ -19,13 +19,14 @@ struct xp_ops {
 	int (*accept_connection)(struct nofuse_queue *ep);
 	struct ep_qe *(*acquire_tag)(struct nofuse_queue *ep,
 				     union nvme_tcp_pdu *pdu,
-				     u16 ccid, u64 pos, u64 len);
-	struct ep_qe *(*get_tag)(struct nofuse_queue *ep, u16 tag);
+				     uint16_t ccid, off_t pos, size_t len);
+	struct ep_qe *(*get_tag)(struct nofuse_queue *ep, uint16_t tag);
 	struct ep_qe *(*get_aen)(struct nofuse_queue *ep);
 	void (*release_tag)(struct nofuse_queue *ep, struct ep_qe *qe);
-	int (*rma_read)(struct nofuse_queue *ep, void *buf, u64 len);
-	int (*rma_write)(struct nofuse_queue *ep, struct ep_qe *qe, u64 len);
-	int (*prep_rma_read)(struct nofuse_queue *ep, u16 ttag);
+	int (*rma_read)(struct nofuse_queue *ep, void *buf, size_t len);
+	int (*rma_write)(struct nofuse_queue *ep, struct ep_qe *qe,
+			 size_t len);
+	int (*prep_rma_read)(struct nofuse_queue *ep, uint16_t ttag);
 	int (*send_rsp)(struct nofuse_queue *ep, struct nvme_completion *comp);
 	int (*read_msg)(struct nofuse_queue *ep);
 	int (*handle_msg)(struct nofuse_queue *ep);
